fix: ajusta responsividade na seção de patrocinadores, move componentes da Home para a pasta  home @thaina-lopes @Jhamyllie
